FIX: Add a lap only on first Start, not subsequent
This fixes a bug that was introduced in 64cd805. A new lap was
being created every time startTimer() was called. This should only
be the functionality on the *first* call (when the app first
launches and after a rest). Every time stopTimer() is called, then
startTimer() is called without a reset, the previously running lap
should update. No new lap should be added.